댓글 작성 구현

+++ skuhub/src/main/java/org/skuhub/skuhub/repository/taxi/TaxiCommentRepository.java
@@ -1,9 +1,13 @@
 package org.skuhub.skuhub.repository.taxi;
 
 import org.skuhub.skuhub.model.taxi.TaxiCommentJpaEntity;
+import org.skuhub.skuhub.model.taxi.TaxiShareJpaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface TaxiCommentRepository extends JpaRepository<TaxiCommentJpaEntity, Integer> {
+public interface TaxiCommentRepository extends JpaRepository<TaxiCommentJpaEntity, Long> {
+//    List<TaxiCommentJpaEntity> findByPostId(TaxiShareJpaEntity postId);
 }
